templates now a folder and move stuff to utils folder
